Rename back to List, remove redundant stories
