+++ src/ProductSwitch/ProductSwitch.stories.tsx
@@ -4,7 +4,7 @@ import { Story, Meta } from '@storybook/react';
 import { ProductSwitch } from './ProductSwitch';
 import { Menu } from '../Menu';
 import { Popover } from '../Popover';
-import { ListGroup } from '../List';
+import { List } from '../List';
 import { ReactComponent as SwitchAppIconColour } from '../Header/switch-app-icon-color.svg';
 import { ItemWrapper, StyledButtonTitle, StyledNewBadge, StyledText, SwitchAppWrapper } from './ProductSwitch.styles';
 import { EdozoLogo2 } from '../EdozoLogo2';
@@ -31,26 +31,26 @@ ProductSwitchWithPopover.args = {
               visible
               content={
                 <div style={{ margin: '10px 0' }}>
-                  <ListGroup variant="platform">
-                    <ListGroup.Item onClick={() => console.log('Do a thing')}>
+                  <List variant="platform">
+                    <List.Item onClick={() => console.log('Do a thing')}>
                       <ItemWrapper>
                         <EdozoLogo2 appName="maps" />
                         <StyledText>Text</StyledText>
                       </ItemWrapper>
-                    </ListGroup.Item>
-                    <ListGroup.Item onClick={() => console.log('Do a thing')}>
+                    </List.Item>
+                    <List.Item onClick={() => console.log('Do a thing')}>
                       <ItemWrapper>
                         <EdozoLogo2 appName="occupiers" />
                         <StyledText>Text</StyledText>
                       </ItemWrapper>
-                    </ListGroup.Item>
-                    <ListGroup.Item onClick={() => console.log('Do a thing')}>
+                    </List.Item>
+                    <List.Item onClick={() => console.log('Do a thing')}>
                       <ItemWrapper>
                         <EdozoLogo2 appName="insight" />
                         <StyledText>Text</StyledText>
                       </ItemWrapper>
-                    </ListGroup.Item>
-                  </ListGroup>
+                    </List.Item>
+                  </List>
                 </div>
               }
             >
